@@ -1,5 +1,5 @@
 package httpx
 
 type Logger interface {
-	Printf(format string, args ...interface{})
+	Printf(format string, args ...any)
 }
