@@ -8,7 +8,7 @@ import (
 	"github.com/ungerik/go-httpx/httperr"
 )
 
-type JSON func(http.ResponseWriter, *http.Request) (response interface{}, err error)
+type JSON func(http.ResponseWriter, *http.Request) (response any, err error)
 
 func (handlerFunc JSON) ServeHTTP(writer http.ResponseWriter, request *http.Request) {
 	if CatchPanics {
@@ -25,7 +25,7 @@ func (handlerFunc JSON) ServeHTTP(writer http.ResponseWriter, request *http.Requ
 	WriteJSON(writer, response)
 }
 
-func WriteJSON(writer http.ResponseWriter, response interface{}) {
+func WriteJSON(writer http.ResponseWriter, response any) {
 	b, err := EncodeJSON(response)
 	if err != nil {
 		httperr.WriteInternalServerError(err, writer)
@@ -35,7 +35,7 @@ func WriteJSON(writer http.ResponseWriter, response interface{}) {
 	writer.Write(b) //#nosec G104
 }
 
-func EncodeJSON(response interface{}) ([]byte, error) {
+func EncodeJSON(response any) ([]byte, error) {
 	if PrettyPrint {
 		return json.MarshalIndent(response, "", PrettyPrintIndent)
 	}
